replaced qvariants and string refs in tree view
This is actually matching the original logic and is much safer and
cleaner, instead of opaque refs, we are using actual class objects

We no longer need to do some weird hacks to retrieve attributes of
objects and instead we can just use the native methods they provide.

This is consistent and guaranteed to work, compile time checked, it
works with every xapi version as well.